+++ crates/ruma-federation-api/CHANGELOG.md
@@ -3,6 +3,7 @@
 Improvements:
 
 * Stabilize support for getting an event by timestamp (MSC3030 / Matrix 1.6)
+* Stabilize support for partial state in `v2/send_join` (MSC3706 / Matrix 1.6)
 
 # 0.7.0